requirements.txt
contourpy==1.1.0
cycler==0.11.0
fonttools==4.42.1
kiwisolver==1.4.5
matplotlib==3.7.2
numpy==1.25.2
packaging==23.1
paho-mqtt==1.6.1
Pillow==10.0.0
pyparsing==3.0.9
PySimpleGUI==4.60.5
python-dateutil==2.8.2
scipy==1.11.2
six==1.16.0
